:building_construction: Update: Backend and Fronted connection has been linked.

+++ CreditCardValidation/src/main/java/com/example/creditcardvalidation/repositories/CardRepository.java
@@ -3,5 +3,6 @@ package com.example.creditcardvalidation.repositories;
 import com.example.creditcardvalidation.models.entities.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card, Integer> {
+public interface CardRepository extends JpaRepository<Card, Long> {
+
 }
